requirements.txt
# Core dependencies
openai>=1.3.0
supabase>=2.5.0
python-dotenv>=1.0.0

# PDF processing
pypdf>=4.0.0

# Text processing
langchain-text-splitters>=0.0.2

# Web framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
websockets>=12.0

# HTTP client
requests>=2.31.0

# Optional: async support
httpx>=0.25.0
